+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/anime/source/browse/BrowseAnimeSourceScreen.kt
@@ -272,6 +272,7 @@ data class BrowseAnimeSourceScreen(
                             BrowseAnimeSourceScreenModel.Dialog.Migrate(dialog.anime, dialog.duplicate),
                         )
                     },
+                    onMerge = onDismissRequest,
                     isManga = false,
                 )
             }