+++ app/src/main/java/eu/kanade/tachiyomi/ui/history/anime/AnimeHistoryTab.kt
@@ -122,6 +122,7 @@ fun Screen.animeHistoryTab(
                         onMigrate = {
                             screenModel.showMigrateDialog(dialog.anime, dialog.duplicate)
                         },
+                        onMerge = onDismissRequest,
                         isManga = false,
                     )
                 }